P15: rebuild deployment onto Singapore Supabase, single-provider hosting
Schema (A): drop 8 unused tables (00006/00007 -> _archive), daily_returns/
index_returns become views over daily_bars/market_index_bars, indicator
columns numeric -> double precision, applied to Supabase Singapore
(qhbfjgheeyckefcwtmcq) through a migration runner (schema_migrations: 11 rows).

Write path (B): upsert.py batches through execute_values (page_size=500)
instead of executemany; universe synced, price history and indicators
loaded, artifact ae2010a4ad426 loaded/activated and verified field-for-field
against disk.

Read path (C): GZipMiddleware on the API; 60s in-process TTL cache plus
Cache-Control on /api/* GET 200s.

Deploy (D1-D3): render.yaml declares both services (API on starter),
same-origin rewrite accepted by validateApiBaseUrl, netlify.toml deleted.
D4 (retire stale Render services) and D5 (delete Sydney Supabase project)
are deliberately not included here - destructive infra actions gated on
the user driving them directly.

One decision record (D-31) bundles the provider consolidation and data-tier
reshape; RUNBOOK and related docs updated to match.

+++ supabase/migrations/00013_market_home_views.sql
@@ -50,9 +50,13 @@
 -- return is legitimately negative, and NULL means "not computed" — which for
 -- ret_252d is every row whose ticker has fewer than 253 loaded sessions behind
 -- it. Existing rows get NULL and stay NULL until the indicator build reruns.
+--
+-- `double precision`, matching the columns it sits beside since P15 changed them
+-- from numeric. A `numeric` here would be the one decimal column in a float8
+-- table, which psycopg2 would hand back as a lone Decimal among floats.
 -- -----------------------------------------------------------------------------
 ALTER TABLE technical_indicators_daily
-    ADD COLUMN IF NOT EXISTS ret_252d numeric;
+    ADD COLUMN IF NOT EXISTS ret_252d double precision;
 
 COMMENT ON COLUMN technical_indicators_daily.ret_252d IS
     'Trailing simple return over 252 SESSIONS (not 365 days), as a fraction. NULL until the '
